+++ rest_framework_social_oauth2/oauth2_grants.py
@@ -9,10 +9,10 @@ from django.core.urlresolvers import reverse
 from oauthlib.oauth2.rfc6749 import errors
 from oauthlib.oauth2.rfc6749.grant_types.refresh_token import RefreshTokenGrant
 
-from social.apps.django_app.views import NAMESPACE
-from social.apps.django_app.utils import load_backend, load_strategy
-from social.exceptions import MissingBackend, SocialAuthBaseException
-from social.utils import requests
+from social_django.views import NAMESPACE
+from social_django.utils import load_backend, load_strategy
+from social_core.exceptions import MissingBackend, SocialAuthBaseException
+from social_core.utils import requests
 
 
 log = logging.getLogger(__name__)