+++ rest_framework_social_oauth2/settings.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
 from django.conf import settings
 
-PROPRIETARY_BACKEND_NAME = getattr(settings, 'PROPRIETARY_BACKEND_NAME', "Django")
+DRFSO2_PROPRIETARY_BACKEND_NAME = getattr(settings, 'DRFSO2_PROPRIETARY_BACKEND_NAME', "Django")
+DRFSO2_URL_NAMESPACE = getattr(settings, 'DRFSO2_URL_NAMESPACE', "")