+++ rest_framework_social_oauth2/views.py
@@ -2,6 +2,8 @@
 import json
 
 from braces.views import CsrfExemptMixin
+from oauthlib.oauth2.rfc6749.endpoints.token import TokenEndpoint
+from oauth2_provider.oauth2_backends import OAuthLibCore
 from oauth2_provider.ext.rest_framework import OAuth2Authentication
 from oauth2_provider.models import Application, AccessToken
 from oauth2_provider.settings import oauth2_settings
@@ -16,7 +18,7 @@ from .oauth2_backends import KeepRequestCore
 from .oauth2_endpoints import SocialTokenServer
 
 
-class ConvertTokenView(CsrfExemptMixin, OAuthLibMixin, APIView):
+class TokenView(CsrfExemptMixin, OAuthLibMixin, APIView):
     """
     Implements an endpoint to provide access tokens
 
@@ -26,13 +28,12 @@ class ConvertTokenView(CsrfExemptMixin, OAuthLibMixin, APIView):
     * Password
     * Client credentials
     """
-    server_class = SocialTokenServer
+    server_class = oauth2_settings.OAUTH2_SERVER_CLASS
     validator_class = oauth2_settings.OAUTH2_VALIDATOR_CLASS
-    oauthlib_backend_class = KeepRequestCore
+    oauthlib_backend_class = OAuthLibCore
     permission_classes = (permissions.AllowAny,)
 
     def post(self, request, *args, **kwargs):
-
         # Use the rest framework `.data` to fake the post body of the django request.
         request._request.POST = request._request.POST.copy()
         for key, value in request.data.items():
@@ -40,6 +41,57 @@ class ConvertTokenView(CsrfExemptMixin, OAuthLibMixin, APIView):
 
         url, headers, body, status = self.create_token_response(request._request)
         response = Response(data=json.loads(body), status=status)
+
+        for k, v in headers.items():
+            response[k] = v
+        return response
+
+
+class ConvertTokenView(CsrfExemptMixin, OAuthLibMixin, APIView):
+    """
+    Implements an endpoint to convert a provider token to an access token
+
+    The endpoint is used in the following flows:
+
+    * Authorization code
+    * Client credentials
+    """
+    server_class = SocialTokenServer
+    validator_class = oauth2_settings.OAUTH2_VALIDATOR_CLASS
+    oauthlib_backend_class = KeepRequestCore
+    permission_classes = (permissions.AllowAny,)
+
+    def post(self, request, *args, **kwargs):
+        # Use the rest framework `.data` to fake the post body of the django request.
+        request._request.POST = request._request.POST.copy()
+        for key, value in request.data.items():
+            request._request.POST[key] = value
+
+        url, headers, body, status = self.create_token_response(request._request)
+        response = Response(data=json.loads(body), status=status)
+
+        for k, v in headers.items():
+            response[k] = v
+        return response
+
+
+class RevokeTokenView(CsrfExemptMixin, OAuthLibMixin, APIView):
+    """
+    Implements an endpoint to revoke access or refresh tokens
+    """
+    server_class = oauth2_settings.OAUTH2_SERVER_CLASS
+    validator_class = oauth2_settings.OAUTH2_VALIDATOR_CLASS
+    oauthlib_backend_class = OAuthLibCore
+    permission_classes = (permissions.AllowAny,)
+
+    def post(self, request, *args, **kwargs):
+        # Use the rest framework `.data` to fake the post body of the django request.
+        request._request.POST = request._request.POST.copy()
+        for key, value in request.data.items():
+            request._request.POST[key] = value
+
+        url, headers, body, status = self.create_revocation_response(request._request)
+        response = Response(data=json.loads(body) if body else '', status=status if body else 204)
 
         for k, v in headers.items():
             response[k] = v